BLD,TST: Split Windows test based on needed feature.
Some tests are for newlines (LF or CRLF?), others for default int
(32-bit on 32-bit platforms or Windows; 64-bit on other 64-bit platforms)
one seems to be checking specifically the compiler.
There are a few checking on actual limitations of the Windows operating system that Cygwin hasn't worked around.
Mostly this is trying to get the number of test failures down.